+++ demos/probabilistic_data_structures/hyperloglog_demo.c
@@ -18,14 +18,16 @@ void hyperloglog_demo(void)
         return;
     }
 
-    printf("[INITIALIZED] Registers: %zu (2^%d), Memory: %zu bytes\n", hll->m, precision, hll->m);
+    printf("[INITIALIZED] Registers: %lu (2^%d), Memory: %lu bytes\n", (unsigned long)hll->m,
+           precision, (unsigned long)hll->m);
 
     size_t unique_count = 2500;
-    printf("Inserting %zu distinct elements into HyperLogLog stream...\n", unique_count);
+    printf("Inserting %lu distinct elements into HyperLogLog stream...\n",
+           (unsigned long)unique_count);
     char buf[64];
     for (size_t i = 0; i < unique_count; i++)
     {
-        snprintf(buf, sizeof(buf), "unique_stream_item_%zu", i);
+        snprintf(buf, sizeof(buf), "unique_stream_item_%lu", (unsigned long)i);
         hll_add(hll, buf, strlen(buf));
     }
 
@@ -33,8 +35,8 @@ void hyperloglog_demo(void)
     double error_pct = (fabs((double)est - (double)unique_count) / (double)unique_count) * 100.0;
 
     printf("\nResults:\n");
-    printf(" Actual Unique Count      : %zu\n", unique_count);
-    printf(" HLL Estimated Count     : %llu\n", (unsigned long long)est);
+    printf(" Actual Unique Count      : %lu\n", (unsigned long)unique_count);
+    printf(" HLL Estimated Count     : %I64u\n", (unsigned long long)est);
     printf(" Relative Estimation Error: %.2f%%\n", error_pct);
 
     hll_free(hll);
